Handle extra com.ibm.smf.format.SMF argument
Since the change to building with Maven in b040e60, the
JAR is now self-executable and specifies the main class
com.ibm.smf.format.SMF meaning that when executing with
-jar, specifying the class name is not needed. However,
a lot of examples still show this sort of execution.
Running with the class name as the first argument will
cause an error because all arguments are expected to have
parentheses. Therefore, this commit allows for an extra
com.ibm.smf.format.SMF argument which is simply ignored.
Things will work the same as before if the user is
running without an executable JAR and instead puts a
packaged JAR on the classpath.

Signed-off-by: Kevin Grigorenko <kevin.grigorenko@us.ibm.com>