Allow multiple INFILEs for com.ibm.smf.format.SMF
It's common for administrators to provide SMF data across multiple
data sets due to size constraints, so this adds the ability to run
the same PLUGIN across multiple INFILE data sets. The change is
pretty small but the diff is relatively big because I could no
longer handle the strange formatting so I performed a format
on the whole files through Eclipse, so there are a lot of whitespace
changes; however, the actual changes are pretty simple:

1. In SMF.main, add all INFILEs to an ArrayList and then iterate
   over that list and call Interpreter.interpret for each.
2. Add an optional boolean parameter to Interpreter.interpret
   to specify whether or not to run processingComplete
   on the filter at the end of the interpret method. This defaults
   to true to match the old behavior (e.g. used by JclSmf).
3. Update SMF.main to pass false to Interpreter.interpret
   and then call Interpreter.performFilterCompletion after
   all files have been processed.

The commit also includes a minor bug fix for the ResponseTimes
plugin when there are non-120 records.

Signed-off-by: Kevin Grigorenko <kevin.grigorenko@us.ibm.com>